Fix crash when trying to render destroyed window (#1)
When a window is hidden (`unmap_notify`), it's can be immediately
followed by destruction (`destroy_notify`) if the window is closing.
Without this extra check, we attempted to render a window that may
already be destroyed by the X server, causing errors when our rendering
commands reach the server.

It's also just better if we don't try to render a hidden window at all.


### Testing strategy

 1. `Xephyr :99 -screen 1920x1080x24 -retro`
 1. Start the compositing manager: `DISPLAY=:99 zig build run-main`
1. Start and stop some other X graphical application on the same display
`DISPLAY=:99 some-app`
 1. Start the other X graphical application again
1. Notice that the compositing manager crashed with the following error
in the logs:
     ```shell
info: RENDER extension: opcode=139 base_event_code=0 base_error_code=142
    ...
error: expected a reply for `x.get_window_attributes` but got
x.ServerMsgTaggedUnion{ .err = x.ServerMsg.Error{ .reponse_type =
x.ErrorKind.err, .code = x.ErrorCode(143), .sequence = 72, .generic =
4194312, .minor_opcode = 8, .major_opcode = x.Opcode(139), .data = { 0,
0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 } } }
    ```
- To break down the `x.ServerMsg.Error`, we have to cross-reference the
`opcode` and `base_error_code` of the extensions. In this case, it lines
up with the `RENDER` extension.
- `.major_opcode = x.Opcode(139)` corresponds to the `RENDER` extension
`opcode`
- `.minor_opcode = 8` corresponds to the [`x.render.ExtOpcode.composite
=
8`](https://github.com/marler8997/zigx/blob/69fef6c194df14361ce85b8f27bb80febf26b6e1/src/render.zig#L52)
because `base_event_code=0`
- `.code = x.ErrorCode(143)` corresponds to [`x.render.ExtOpcode.Picture
=
1`](https://github.com/marler8997/zigx/blob/69fef6c194df14361ce85b8f27bb80febf26b6e1/src/render.zig#L91)
because `base_error_code=142`
- `.generic = 4194312` points to the picture ID that does not exist
anymore
- It would be nice if we could catch our errors right where we sent the
request in the first place. The problem is not actually with
`x.get_window_attributes` but a previous `x.render.composite` request
that is "asynchronous" and has no reply but can return an error. The
problem is that not all requests have replies to listen for (they are
considered "asynchronous"). And we can't just spend time listening for
the absence of an error. We need some sort of ack. "[If] the request has
no reply (it is asynchronous), although errors may still be reported"
([source](https://www.x.org/releases/current/doc/xproto/x11protocol.html#Syntactic_Conventions)).
Perhaps if requests are processed sequentially and causality is
preserved, we could use some dummy ping-pong request/reply and we either
see the reply or the error from the previous async request.
- It would also be way better if parsed the errors to automatically
format them with all of the plain-english to point to what extension had
the problem, which request, which error, etc instead of needing to
cross-reference everything.